factor out event code
